requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.23
tqdm>=4.65
python-chess>=1.999
